Move `initialize_result_object` to result.py
